Fixed interface and implementation name
